requirements.txt
Django==1.9.5
django-simple-captcha==0.5.1
django-tinymce==2.2.0
django-widget-tweaks==1.4.1
Pillow==3.1.0
six==1.10.0
wheel==0.24.0
MySQL-python
pytz
